move functions outside of the file resolver
