+++ core/workflow-operator/src/main/scala/edu/uci/ics/amber/operator/source/cache/CacheSourceOpExec.scala
@@ -2,8 +2,7 @@ package edu.uci.ics.amber.operator.source.cache
 
 import com.typesafe.scalalogging.LazyLogging
 import edu.uci.ics.amber.core.executor.SourceOperatorExecutor
-import edu.uci.ics.amber.core.storage.DocumentFactory
-import edu.uci.ics.amber.core.storage.FileResolver.decodeVFSUri
+import edu.uci.ics.amber.core.storage.{DocumentFactory, VFSURIFactory}
 import edu.uci.ics.amber.core.storage.VFSResourceType.MATERIALIZED_RESULT
 import edu.uci.ics.amber.core.storage.model.VirtualDocument
 import edu.uci.ics.amber.core.tuple.{Tuple, TupleLike}
@@ -11,7 +10,7 @@ import edu.uci.ics.amber.core.tuple.{Tuple, TupleLike}
 import java.net.URI
 
 class CacheSourceOpExec(storageUri: URI) extends SourceOperatorExecutor with LazyLogging {
-  val (_, _, _, _, resourceType) = decodeVFSUri(storageUri)
+  val (_, _, _, _, resourceType) = VFSURIFactory.decodeVFSUri(storageUri)
   if (resourceType != MATERIALIZED_RESULT) {
     throw new RuntimeException("The storage URI must point to an materialized result storage")
   }
